+++ src/library/ContentDescriptor/definition.ts
